Misc. ExtUI LCD cleanup
