+++ Marlin/src/lcd/extui/anycubic_chiron/chiron_extui.cpp
@@ -37,17 +37,17 @@ using namespace Anycubic;
 
 namespace ExtUI {
 
-  void onStartup() { Chiron.Startup(); }
+  void onStartup() { chiron.startup(); }
 
-  void onIdle() { Chiron.IdleLoop(); }
+  void onIdle() { chiron.idleLoop(); }
 
   void onPrinterKilled(FSTR_P const error, FSTR_P const component) {
-    Chiron.PrinterKilled(error, component);
+    chiron.printerKilled(error, component);
   }
 
-  void onMediaInserted() { Chiron.MediaEvent(AC_media_inserted); }
-  void onMediaError()    { Chiron.MediaEvent(AC_media_error);    }
-  void onMediaRemoved()  { Chiron.MediaEvent(AC_media_removed);  }
+  void onMediaInserted() { chiron.mediaEvent(AC_media_inserted); }
+  void onMediaError()    { chiron.mediaEvent(AC_media_error);    }
+  void onMediaRemoved()  { chiron.mediaEvent(AC_media_removed);  }
 
   void onPlayTone(const uint16_t frequency, const uint16_t duration) {
     #if ENABLED(SPEAKER)
@@ -55,15 +55,15 @@ namespace ExtUI {
     #endif
   }
 
-  void onPrintTimerStarted() { Chiron.TimerEvent(AC_timer_started); }
-  void onPrintTimerPaused()  { Chiron.TimerEvent(AC_timer_paused);  }
-  void onPrintTimerStopped() { Chiron.TimerEvent(AC_timer_stopped); }
+  void onPrintTimerStarted() { chiron.timerEvent(AC_timer_started); }
+  void onPrintTimerPaused()  { chiron.timerEvent(AC_timer_paused);  }
+  void onPrintTimerStopped() { chiron.timerEvent(AC_timer_stopped); }
   void onPrintDone() {}
 
-  void onFilamentRunout(const extruder_t)            { Chiron.FilamentRunout();             }
+  void onFilamentRunout(const extruder_t)            { chiron.filamentRunout();             }
 
-  void onUserConfirmRequired(const char * const msg) { Chiron.ConfirmationRequest(msg);     }
-  void onStatusChanged(const char * const msg)       { Chiron.StatusChange(msg);            }
+  void onUserConfirmRequired(const char * const msg) { chiron.confirmationRequest(msg);     }
+  void onStatusChanged(const char * const msg)       { chiron.statusChange(msg);            }
 
   void onHomingStart() {}
   void onHomingDone() {}
@@ -127,7 +127,7 @@ namespace ExtUI {
       // Called when power-loss state is detected
     }
     // Called on resume from power-loss
-    void onPowerLossResume() { Chiron.PowerLossRecovery(); }
+    void onPowerLossResume() { chiron.powerLossRecovery(); }
   #endif
 
   #if HAS_PID_HEATING
