+++ Marlin/src/lcd/extui/anycubic_i3mega/anycubic_extui.cpp
@@ -35,20 +35,20 @@
 
 namespace ExtUI {
 
-  void onStartup()        { AnycubicTFT.OnSetup(); }
-  void onIdle()           { AnycubicTFT.OnCommandScan(); }
-  void onPrinterKilled(FSTR_P const error, FSTR_P const component) { AnycubicTFT.OnKillTFT(); }
-  void onMediaInserted()  { AnycubicTFT.OnSDCardStateChange(true); }
-  void onMediaError()     { AnycubicTFT.OnSDCardError(); }
-  void onMediaRemoved()   { AnycubicTFT.OnSDCardStateChange(false); }
+  void onStartup()        { anycubicTFT.onSetup(); }
+  void onIdle()           { anycubicTFT.onCommandScan(); }
+  void onPrinterKilled(FSTR_P const error, FSTR_P const component) { anycubicTFT.onKillTFT(); }
+  void onMediaInserted()  { anycubicTFT.onSDCardStateChange(true); }
+  void onMediaError()     { anycubicTFT.onSDCardError(); }
+  void onMediaRemoved()   { anycubicTFT.onSDCardStateChange(false); }
   void onPlayTone(const uint16_t frequency, const uint16_t duration) {
     TERN_(SPEAKER, ::tone(BEEPER_PIN, frequency, duration));
   }
-  void onPrintTimerStarted()  { AnycubicTFT.OnPrintTimerStarted(); }
-  void onPrintTimerPaused()   { AnycubicTFT.OnPrintTimerPaused(); }
-  void onPrintTimerStopped()  { AnycubicTFT.OnPrintTimerStopped(); }
-  void onFilamentRunout(const extruder_t extruder)   { AnycubicTFT.OnFilamentRunout(); }
-  void onUserConfirmRequired(const char * const msg) { AnycubicTFT.OnUserConfirmRequired(msg); }
+  void onPrintTimerStarted()  { anycubicTFT.onPrintTimerStarted(); }
+  void onPrintTimerPaused()   { anycubicTFT.onPrintTimerPaused(); }
+  void onPrintTimerStopped()  { anycubicTFT.onPrintTimerStopped(); }
+  void onFilamentRunout(const extruder_t extruder)   { anycubicTFT.onFilamentRunout(); }
+  void onUserConfirmRequired(const char * const msg) { anycubicTFT.onUserConfirmRequired(msg); }
   void onStatusChanged(const char * const msg) {}
 
   void onHomingStart() {}
