+++ Marlin/src/lcd/extui/anycubic_i3mega/anycubic_i3mega_lcd.cpp
@@ -24,6 +24,8 @@
 #if ENABLED(ANYCUBIC_LCD_I3MEGA)
 
 //#define ANYCUBIC_LCD_DEBUG
+#define DEBUG_OUT ANYCUBIC_LCD_DEBUG
+#include "../../../core/debug_out.h"
 
 #include "anycubic_i3mega_lcd.h"
 #include "../ui_api.h"
@@ -40,13 +42,8 @@
 #define SEND_PGM_VAL(x,y) (send_P(PSTR(x)), sendLine(i16tostr3rj(y)))
 #define SEND(x)           send(x)
 #define SENDLINE(x)       sendLine(x)
-#if ENABLED(ANYCUBIC_LCD_DEBUG)
-  #define SENDLINE_DBG_PGM(x,y)       do{ sendLine_P(PSTR(x)); SERIAL_ECHOLNPGM(y); }while(0)
-  #define SENDLINE_DBG_PGM_VAL(x,y,z) do{ sendLine_P(PSTR(x)); SERIAL_ECHOLNPGM(y, z); }while(0)
-#else
-  #define SENDLINE_DBG_PGM(x,y)       sendLine_P(PSTR(x))
-  #define SENDLINE_DBG_PGM_VAL(x,y,z) sendLine_P(PSTR(x))
-#endif
+#define SENDLINE_DBG_PGM(x,y)       do{ sendLine_P(PSTR(x)); DEBUG_ECHOLNPGM(y); }while(0)
+#define SENDLINE_DBG_PGM_VAL(x,y,z) do{ sendLine_P(PSTR(x)); DEBUG_ECHOLNPGM(y, z); }while(0)
 
 // Append ".gcode" to filename, if requested. Used for some DGUS-clone displays with built-in filter.
 // Filenames are limited to 26 characters, so the actual name for the FILENAME can be 20 characters at most.
@@ -54,21 +51,20 @@
 #define SPECIAL_MENU_FILENAME(A) A TERN_(ANYCUBIC_LCD_GCODE_EXT, ".gcode")
 #define SPECIAL_MENU_ALTNAME(A, B) TERN(ANYCUBIC_LCD_GCODE_EXT, A ".gcode", B)
 
-AnycubicTFTClass AnycubicTFT;
+AnycubicTFTClass anycubicTFT;
 
-char AnycubicTFTClass::TFTcmdbuffer[TFTBUFSIZE][TFT_MAX_CMD_SIZE];
-int AnycubicTFTClass::TFTbuflen = 0,
-    AnycubicTFTClass::TFTbufindr = 0,
-    AnycubicTFTClass::TFTbufindw = 0;
+char AnycubicTFTClass::tftCommands[TFTBUFSIZE][TFT_MAX_CMD_SIZE];
+int AnycubicTFTClass::tftBufLen = 0,
+    AnycubicTFTClass::tftBufIndR = 0,
+    AnycubicTFTClass::tftBufIndW = 0;
 char AnycubicTFTClass::serial3_char;
-int AnycubicTFTClass::serial3_count = 0;
-char* AnycubicTFTClass::TFTstrchr_pointer;
-uint8_t AnycubicTFTClass::SpecialMenu = false;
+char* AnycubicTFTClass::tftStrchrPtr;
+uint8_t AnycubicTFTClass::specialMenu = false;
 AnycubicMediaPrintState AnycubicTFTClass::mediaPrintingState = AMPRINTSTATE_NOT_PRINTING;
 AnycubicMediaPauseState AnycubicTFTClass::mediaPauseState = AMPAUSESTATE_NOT_PAUSED;
 
-char AnycubicTFTClass::SelectedDirectory[30];
-char AnycubicTFTClass::SelectedFile[FILENAME_LENGTH];
+char AnycubicTFTClass::selectedDirectory[30];
+char AnycubicTFTClass::selectedFile[FILENAME_LENGTH];
 
 // Serial helpers
 static void sendNewLine() { LCD_SERIAL.write('\r'); LCD_SERIAL.write('\n'); }
@@ -84,7 +80,7 @@ using namespace ExtUI;
 
 AnycubicTFTClass::AnycubicTFTClass() {}
 
-void AnycubicTFTClass::OnSetup() {
+void AnycubicTFTClass::onSetup() {
   #ifndef LCD_BAUDRATE
     #define LCD_BAUDRATE 115200
   #endif
@@ -94,9 +90,6 @@ void AnycubicTFTClass::OnSetup() {
   delay_ms(10);
 
   // Init the state of the key pins running on the TFT
-  #if BOTH(HAS_MEDIA, HAS_SD_DETECT)
-    SET_INPUT_PULLUP(SD_DETECT_PIN);
-  #endif
   #if ENABLED(FILAMENT_RUNOUT_SENSOR)
     SET_INPUT_PULLUP(FIL_RUNOUT1_PIN);
   #endif
@@ -104,30 +97,26 @@ void AnycubicTFTClass::OnSetup() {
   mediaPrintingState = AMPRINTSTATE_NOT_PRINTING;
   mediaPauseState = AMPAUSESTATE_NOT_PAUSED;
 
-  // DoSDCardStateCheck();
+  // doSDCardStateCheck();
   SENDLINE_DBG_PGM("J12", "TFT Serial Debug: Ready... J12"); // J12 Ready
   delay_ms(10);
 
-  DoFilamentRunoutCheck();
-  SelectedFile[0] = 0;
+  doFilamentRunoutCheck();
+  selectedFile[0] = 0;
 
   #if ENABLED(STARTUP_CHIME)
     injectCommands(F("M300 P250 S554\nM300 P250 S554\nM300 P250 S740\nM300 P250 S554\nM300 P250 S740\nM300 P250 S554\nM300 P500 S831"));
   #endif
-  #if ENABLED(ANYCUBIC_LCD_DEBUG)
-    SERIAL_ECHOLNPGM("TFT Serial Debug: Finished startup");
-  #endif
+  DEBUG_ECHOLNPGM("TFT Serial Debug: Finished startup");
 }
 
-void AnycubicTFTClass::OnCommandScan() {
+void AnycubicTFTClass::onCommandScan() {
   static millis_t nextStopCheck = 0; // used to slow the stopped print check down to reasonable times
   const millis_t ms = millis();
   if (ELAPSED(ms, nextStopCheck)) {
     nextStopCheck = ms + 1000UL;
-    if (mediaPrintingState == AMPRINTSTATE_STOP_REQUESTED && IsNozzleHomed()) {
-      #if ENABLED(ANYCUBIC_LCD_DEBUG)
-        SERIAL_ECHOLNPGM("TFT Serial Debug: Finished stopping print, releasing motors ...");
-      #endif
+    if (mediaPrintingState == AMPRINTSTATE_STOP_REQUESTED && isNozzleHomed()) {
+      DEBUG_ECHOLNPGM("TFT Serial Debug: Finished stopping print, releasing motors ...");
       mediaPrintingState = AMPRINTSTATE_NOT_PRINTING;
       mediaPauseState = AMPAUSESTATE_NOT_PAUSED;
       injectCommands(F("M84\nM27")); // disable stepper motors and force report of SD status
@@ -137,44 +126,36 @@ void AnycubicTFTClass::OnCommandScan() {
     }
   }
 
-  if (TFTbuflen < (TFTBUFSIZE - 1))
-    GetCommandFromTFT();
+  if (tftBufLen < (TFTBUFSIZE) - 1)
+    getCommandFromTFT();
 
-  if (TFTbuflen) {
-    TFTbuflen  = (TFTbuflen - 1);
-    TFTbufindr = (TFTbufindr + 1) % TFTBUFSIZE;
+  if (tftBufLen) {
+    --tftBufLen;
+    tftBufIndR = (tftBufIndR + 1) % (TFTBUFSIZE);
   }
 }
 
-void AnycubicTFTClass::OnKillTFT() {
+void AnycubicTFTClass::onKillTFT() {
   SENDLINE_DBG_PGM("J11", "TFT Serial Debug: Kill command... J11");
 }
 
-void AnycubicTFTClass::OnSDCardStateChange(bool isInserted) {
-  #if ENABLED(ANYCUBIC_LCD_DEBUG)
-    SERIAL_ECHOLNPGM("TFT Serial Debug: OnSDCardStateChange event triggered...", isInserted);
-  #endif
-  DoSDCardStateCheck();
+void AnycubicTFTClass::onSDCardStateChange(bool isInserted) {
+  DEBUG_ECHOLNPGM("TFT Serial Debug: onSDCardStateChange event triggered...", isInserted);
+  doSDCardStateCheck();
 }
 
-void AnycubicTFTClass::OnSDCardError() {
-  #if ENABLED(ANYCUBIC_LCD_DEBUG)
-    SERIAL_ECHOLNPGM("TFT Serial Debug: OnSDCardError event triggered...");
-  #endif
+void AnycubicTFTClass::onSDCardError() {
+  DEBUG_ECHOLNPGM("TFT Serial Debug: onSDCardError event triggered...");
   SENDLINE_DBG_PGM("J21", "TFT Serial Debug: On SD Card Error ... J21");
 }
 
-void AnycubicTFTClass::OnFilamentRunout() {
-  #if ENABLED(ANYCUBIC_LCD_DEBUG)
-    SERIAL_ECHOLNPGM("TFT Serial Debug: FilamentRunout triggered...");
-  #endif
-  DoFilamentRunoutCheck();
+void AnycubicTFTClass::onFilamentRunout() {
+  DEBUG_ECHOLNPGM("TFT Serial Debug: onFilamentRunout triggered...");
+  doFilamentRunoutCheck();
 }
 
-void AnycubicTFTClass::OnUserConfirmRequired(const char * const msg) {
-  #if ENABLED(ANYCUBIC_LCD_DEBUG)
-    SERIAL_ECHOLNPGM("TFT Serial Debug: OnUserConfirmRequired triggered... ", msg);
-  #endif
+void AnycubicTFTClass::onUserConfirmRequired(const char * const msg) {
+  DEBUG_ECHOLNPGM("TFT Serial Debug: onUserConfirmRequired triggered... ", msg);
 
   #if HAS_MEDIA
     /**
@@ -225,190 +206,188 @@ void AnycubicTFTClass::OnUserConfirmRequired(const char * const msg) {
   #endif
 }
 
-float AnycubicTFTClass::CodeValue() {
-  return (strtod(&TFTcmdbuffer[TFTbufindr][TFTstrchr_pointer - TFTcmdbuffer[TFTbufindr] + 1], nullptr));
+float AnycubicTFTClass::codeValue() {
+  return (strtod(&tftCommands[tftBufIndR][tftStrchrPtr - tftCommands[tftBufIndR] + 1], nullptr));
 }
 
-bool AnycubicTFTClass::CodeSeen(char code) {
-  TFTstrchr_pointer = strchr(TFTcmdbuffer[TFTbufindr], code);
-  return !!TFTstrchr_pointer; // Return True if a character was found
+bool AnycubicTFTClass::codeSeen(char code) {
+  tftStrchrPtr = strchr(tftCommands[tftBufIndR], code);
+  return !!tftStrchrPtr; // Return True if a character was found
 }
 
-bool AnycubicTFTClass::IsNozzleHomed() {
+bool AnycubicTFTClass::isNozzleHomed() {
   const float xPosition = getAxisPosition_mm((axis_t) X);
   const float yPosition = getAxisPosition_mm((axis_t) Y);
   return WITHIN(xPosition, X_MIN_POS - 0.1, X_MIN_POS + 0.1) &&
          WITHIN(yPosition, Y_MIN_POS - 0.1, Y_MIN_POS + 0.1);
 }
 
-void AnycubicTFTClass::HandleSpecialMenu() {
+void AnycubicTFTClass::handleSpecialMenu() {
   /**
    * NOTE: that the file selection command actual lowercases the entire selected file/foldername, so charracter comparisons need to be lowercase.
    */
-  if (SelectedDirectory[0] == '<') {
-    switch (SelectedDirectory[1]) {
-      case 'e': // "<exit>"
-        SpecialMenu = false;
-        return;
+  if (selectedDirectory[0] != '<') {
+    DEBUG_ECHOLNPGM("TFT Serial Debug: Attempted to handleSpecialMenu on non-special menu... ", selectedDirectory);
+    return;
+  }
+
+  switch (selectedDirectory[1]) {
+    default: break;
+
+    case 'e': specialMenu = false; return; // "<exit>"
+
+    #if ENABLED(PROBE_MANUALLY)
+
+      case '0':
+        switch (selectedDirectory[2]) {
+          default: break;
+
+          case '1': // "<01ZUp0.1>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Z Up 0.1"));
+            injectCommands(F("G91\nG1 Z+0.1\nG90"));
+            break;
+
+          case '2': // "<02ZUp0.02>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Z Up 0.02"));
+            injectCommands(F("G91\nG1 Z+0.02\nG90"));
+            break;
+
+          case '3': // "<03ZDn0.02>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Z Down 0.02"));
+            injectCommands(F("G91\nG1 Z-0.02\nG90"));
+            break;
+
+          case '4': // "<04ZDn0.1>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Z Down 0.1"));
+            injectCommands(F("G91\nG1 Z-0.1\nG90"));
+            break;
+
+          case '5': // "<05PrehtBed>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Preheat Bed"));
+            injectCommands(F("M140 S65"));
+            break;
+
+          case '6': // "<06SMeshLvl>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Start Mesh Leveling"));
+            injectCommands(F("G29S1"));
+            break;
+
+          case '7': // "<07MeshNPnt>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Next Mesh Point"));
+            injectCommands(F("G29S2"));
+            break;
+
+          case '8': // "<08HtEndPID>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Auto Tune Hotend PID"));
+            // need to dwell for half a second to give the fan a chance to start before the pid tuning starts
+            injectCommands(F("M106 S204\nG4 P500\nM303 E0 S215 C15 U1"));
+            break;
+
+          case '9': // "<09HtBedPID>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Auto Tune Hotbed Pid"));
+            injectCommands(F("M303 E-1 S65 C6 U1"));
+            break;
+        }
         break;
 
-        #if ENABLED(PROBE_MANUALLY)
-          case '0':
-            switch (SelectedDirectory[2]) {
-              case '1': // "<01ZUp0.1>"
-                SERIAL_ECHOLNPGM("Special Menu: Z Up 0.1");
-                injectCommands(F("G91\nG1 Z+0.1\nG90"));
-                break;
+      case '1':
+        switch (selectedDirectory[2]) {
+          default: break;
 
-              case '2': // "<02ZUp0.02>"
-                SERIAL_ECHOLNPGM("Special Menu: Z Up 0.02");
-                injectCommands(F("G91\nG1 Z+0.02\nG90"));
-                break;
-
-              case '3': // "<03ZDn0.02>"
-                SERIAL_ECHOLNPGM("Special Menu: Z Down 0.02");
-                injectCommands(F("G91\nG1 Z-0.02\nG90"));
-                break;
-
-              case '4': // "<04ZDn0.1>"
-                SERIAL_ECHOLNPGM("Special Menu: Z Down 0.1");
-                injectCommands(F("G91\nG1 Z-0.1\nG90"));
-                break;
-
-              case '5': // "<05PrehtBed>"
-                SERIAL_ECHOLNPGM("Special Menu: Preheat Bed");
-                injectCommands(F("M140 S65"));
-                break;
-
-              case '6': // "<06SMeshLvl>"
-                SERIAL_ECHOLNPGM("Special Menu: Start Mesh Leveling");
-                injectCommands(F("G29S1"));
-                break;
-
-              case '7': // "<07MeshNPnt>"
-                SERIAL_ECHOLNPGM("Special Menu: Next Mesh Point");
-                injectCommands(F("G29S2"));
-                break;
-
-              case '8': // "<08HtEndPID>"
-                SERIAL_ECHOLNPGM("Special Menu: Auto Tune Hotend PID");
-                // need to dwell for half a second to give the fan a chance to start before the pid tuning starts
-                injectCommands(F("M106 S204\nG4 P500\nM303 E0 S215 C15 U1"));
-                break;
-
-              case '9': // "<09HtBedPID>"
-                SERIAL_ECHOLNPGM("Special Menu: Auto Tune Hotbed Pid");
-                injectCommands(F("M303 E-1 S65 C6 U1"));
-                break;
-
-              default:
-                break;
-            }
+          case '0': // "<10FWDeflts>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Load FW Defaults"));
+            injectCommands(F("M502\nM300 P105 S1661\nM300 P210 S1108"));
             break;
 
-          case '1':
-            switch (SelectedDirectory[2]) {
-              case '0': // "<10FWDeflts>"
-                SERIAL_ECHOLNPGM("Special Menu: Load FW Defaults");
-                injectCommands(F("M502\nM300 P105 S1661\nM300 P210 S1108"));
-                break;
-
-              case '1': // "<11SvEEPROM>"
-                SERIAL_ECHOLNPGM("Special Menu: Save EEPROM");
-                injectCommands(F("M500\nM300 P105 S1108\nM300 P210 S1661"));
-                break;
-
-              default:
-                break;
-            }
+          case '1': // "<11SvEEPROM>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Save EEPROM"));
+            injectCommands(F("M500\nM300 P105 S1108\nM300 P210 S1661"));
             break;
-        #else // if ENABLED(PROBE_MANUALLY)
-          case '0':
-            switch (SelectedDirectory[2]) {
-              case '1': // "<01PrehtBed>"
-                SERIAL_ECHOLNPGM("Special Menu: Preheat Bed");
-                injectCommands(F("M140 S65"));
-                break;
+        }
+        break;
 
-              case '2': // "<02ABL>"
-                SERIAL_ECHOLNPGM("Special Menu: Auto Bed Leveling");
-                injectCommands(F("G29N"));
-                break;
+    #else // !PROBE_MANUALLY
 
-              case '3': // "<03HtendPID>"
-                SERIAL_ECHOLNPGM("Special Menu: Auto Tune Hotend PID");
-                // need to dwell for half a second to give the fan a chance to start before the pid tuning starts
-                injectCommands(F("M106 S204\nG4 P500\nM303 E0 S215 C15 U1"));
-                break;
+      case '0':
+        switch (selectedDirectory[2]) {
+          default: break;
 
-              case '4': // "<04HtbedPID>"
-                SERIAL_ECHOLNPGM("Special Menu: Auto Tune Hotbed Pid");
-                injectCommands(F("M303 E-1 S65 C6 U1"));
-                break;
-
-              case '5': // "<05FWDeflts>"
-                SERIAL_ECHOLNPGM("Special Menu: Load FW Defaults");
-                injectCommands(F("M502\nM300 P105 S1661\nM300 P210 S1108"));
-                break;
-
-              case '6': // "<06SvEEPROM>"
-                SERIAL_ECHOLNPGM("Special Menu: Save EEPROM");
-                injectCommands(F("M500\nM300 P105 S1108\nM300 P210 S1661"));
-                break;
-
-              case '7': // <07SendM108>
-                SERIAL_ECHOLNPGM("Special Menu: Send User Confirmation");
-                injectCommands(F("M108"));
-                break;
-
-              default:
-                break;
-            }
+          case '1': // "<01PrehtBed>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Preheat Bed"));
+            injectCommands(F("M140 S65"));
             break;
-            #endif  // PROBE_MANUALLY
 
-          default:
+          case '2': // "<02ABL>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Auto Bed Leveling"));
+            injectCommands(F("G29N"));
             break;
-    }
-    #if ENABLED(ANYCUBIC_LCD_DEBUG)
+
+          case '3': // "<03HtendPID>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Auto Tune Hotend PID"));
+            // need to dwell for half a second to give the fan a chance to start before the pid tuning starts
+            injectCommands(F("M106 S204\nG4 P500\nM303 E0 S215 C15 U1"));
+            break;
+
+          case '4': // "<04HtbedPID>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Auto Tune Hotbed Pid"));
+            injectCommands(F("M303 E-1 S65 C6 U1"));
+            break;
+
+          case '5': // "<05FWDeflts>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Load FW Defaults"));
+            injectCommands(F("M502\nM300 P105 S1661\nM300 P210 S1108"));
+            break;
+
+          case '6': // "<06SvEEPROM>"
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Save EEPROM"));
+            injectCommands(F("M500\nM300 P105 S1108\nM300 P210 S1661"));
+            break;
+
+          case '7': // <07SendM108>
+            SERIAL_ECHOLNPGM("Special Menu: ", F("Send User Confirmation"));
+            injectCommands(F("M108"));
+            break;
+        }
+        break;
+
+    #endif // !PROBE_MANUALLY
   }
-  else {
-    SERIAL_ECHOPGM("TFT Serial Debug: Attempted to HandleSpecialMenu on non-special menu... ");
-    SERIAL_ECHOLN(SelectedDirectory);
-    #endif
-  }
+
 }
 
-void AnycubicTFTClass::RenderCurrentFileList() {
+void AnycubicTFTClass::renderCurrentFileList() {
   #if HAS_MEDIA
     uint16_t selectedNumber = 0;
-    SelectedDirectory[0] = 0;
-    SelectedFile[0] = 0;
+    selectedDirectory[0] = 0;
+    selectedFile[0] = 0;
     FileList currentFileList;
 
     SENDLINE_PGM("FN "); // Filelist start
 
-    if (!isMediaInserted() && !SpecialMenu) {
+    if (!isMediaInserted() && !specialMenu) {
       SENDLINE_DBG_PGM("J02", "TFT Serial Debug: No SD Card mounted to render Current File List... J02");
 
       SENDLINE_PGM("<SPECI~1.GCO");
       SENDLINE_PGM(SPECIAL_MENU_FILENAME("<Special Menu>"));
     }
     else {
-      if (CodeSeen('S'))
-        selectedNumber = CodeValue();
+      if (codeSeen('S'))
+        selectedNumber = codeValue();
 
-      if (SpecialMenu)
-        RenderSpecialMenu(selectedNumber);
+      if (specialMenu)
+        renderSpecialMenu(selectedNumber);
       else if (selectedNumber <= currentFileList.count())
-        RenderCurrentFolder(selectedNumber);
+        renderCurrentFolder(selectedNumber);
     }
     SENDLINE_PGM("END"); // Filelist stop
   #endif // HAS_MEDIA
 }
 
-void AnycubicTFTClass::RenderSpecialMenu(uint16_t selectedNumber) {
+void AnycubicTFTClass::renderSpecialMenu(uint16_t selectedNumber) {
+
   switch (selectedNumber) {
+    default: break;
+
     #if ENABLED(PROBE_MANUALLY)
       case 0: // First Page
         SENDLINE_PGM("<01ZUP~1.GCO");
@@ -442,7 +421,9 @@ void AnycubicTFTClass::RenderSpecialMenu(uint16_t selectedNumber) {
         SENDLINE_PGM("<EXIT_~1.GCO");
         SENDLINE_PGM(SPECIAL_MENU_FILENAME("<Exit>"));
         break;
-    #else
+
+    #else // !PROBE_MANUALLY
+
       case 0: // First Page
         SENDLINE_PGM("<01PRE~1.GCO");
         SENDLINE_PGM(SPECIAL_MENU_FILENAME("<Preheat Bed>"));
@@ -465,25 +446,16 @@ void AnycubicTFTClass::RenderSpecialMenu(uint16_t selectedNumber) {
         SENDLINE_PGM(SPECIAL_MENU_FILENAME("<Exit>"));
         break;
 
-        #endif // PROBE_MANUALLY
-
-      default:
-        break;
+    #endif // !PROBE_MANUALLY
   }
 }
 
-void AnycubicTFTClass::RenderCurrentFolder(uint16_t selectedNumber) {
+void AnycubicTFTClass::renderCurrentFolder(uint16_t selectedNumber) {
   FileList currentFileList;
-  uint16_t cnt = selectedNumber;
-  uint16_t max_files;
-  uint16_t dir_files = currentFileList.count();
+  const uint16_t dir_files = currentFileList.count(),
+                 max_files = (dir_files - selectedNumber) < 4 ? dir_files : selectedNumber + 3;
 
-  if ((dir_files - selectedNumber) < 4)
-    max_files = dir_files;
-  else
-    max_files = selectedNumber + 3;
-
-  for (cnt = selectedNumber; cnt <= max_files; cnt++) {
+  for (uint16_t cnt = selectedNumber; cnt <= max_files; cnt++) {
     if (cnt == 0) { // Special Entry
       if (currentFileList.isAtRootDir()) {
         SENDLINE_PGM("<SPECI~1.GCO");
@@ -497,9 +469,7 @@ void AnycubicTFTClass::RenderCurrentFolder(uint16_t selectedNumber) {
     else {
       currentFileList.seek(cnt - 1, false);
 
-      #if ENABLED(ANYCUBIC_LCD_DEBUG)
-        SERIAL_ECHOLN(currentFileList.filename());
-      #endif
+      DEBUG_ECHOLN(currentFileList.filename());
       if (currentFileList.isDir()) {
         SEND_PGM("/");
         SENDLINE(currentFileList.shortFilename());
@@ -515,15 +485,14 @@ void AnycubicTFTClass::RenderCurrentFolder(uint16_t selectedNumber) {
   }
 }
 
-void AnycubicTFTClass::OnPrintTimerStarted() {
+void AnycubicTFTClass::onPrintTimerStarted() {
   #if HAS_MEDIA
     if (mediaPrintingState == AMPRINTSTATE_PRINTING)
       SENDLINE_DBG_PGM("J04", "TFT Serial Debug: Starting SD Print... J04"); // J04 Starting Print
-
   #endif
 }
 
-void AnycubicTFTClass::OnPrintTimerPaused() {
+void AnycubicTFTClass::onPrintTimerPaused() {
   #if HAS_MEDIA
     if (isPrintingFromMedia()) {
       mediaPrintingState = AMPRINTSTATE_PAUSED;
@@ -532,7 +501,7 @@ void AnycubicTFTClass::OnPrintTimerPaused() {
   #endif
 }
 
-void AnycubicTFTClass::OnPrintTimerStopped() {
+void AnycubicTFTClass::onPrintTimerStopped() {
   #if HAS_MEDIA
     if (mediaPrintingState == AMPRINTSTATE_PRINTING) {
       mediaPrintingState = AMPRINTSTATE_NOT_PRINTING;
@@ -545,385 +514,377 @@ void AnycubicTFTClass::OnPrintTimerStopped() {
 
 #define ROUND(val) int((val)+0.5f)
 
-void AnycubicTFTClass::GetCommandFromTFT() {
+void AnycubicTFTClass::getCommandFromTFT() {
+  static int serial_count = 0;
+
   char *starpos = nullptr;
-  while (LCD_SERIAL.available() > 0  && TFTbuflen < TFTBUFSIZE) {
-    serial3_char = LCD_SERIAL.read();
-    if (serial3_char == '\n' ||
-        serial3_char == '\r' ||
-        serial3_char == ':'  ||
-        serial3_count >= (TFT_MAX_CMD_SIZE - 1)
-    ) {
+  while (LCD_SERIAL.available() > 0 && tftBufLen < (TFTBUFSIZE)) {
+    char c = LCD_SERIAL.read();
+    if (c != '\n' && c != '\r' && c != ':' && serial_count < (TFT_MAX_CMD_SIZE - 1)) {
+      tftCommands[tftBufIndW][serial_count++] = c;
+      continue;
+    }
 
-      if (!serial3_count) return; // if empty line
+    if (!serial_count) return; // if empty line
 
-      TFTcmdbuffer[TFTbufindw][serial3_count] = 0; // terminate string
+    tftCommands[tftBufIndW][serial_count] = 0; // terminate string
 
-      if ((strchr(TFTcmdbuffer[TFTbufindw], 'A') != nullptr)) {
-        int16_t a_command;
-        TFTstrchr_pointer = strchr(TFTcmdbuffer[TFTbufindw], 'A');
-        a_command = ((int)((strtod(&TFTcmdbuffer[TFTbufindw][TFTstrchr_pointer - TFTcmdbuffer[TFTbufindw] + 1], nullptr))));
+    if ((strchr(tftCommands[tftBufIndW], 'A') != nullptr)) {
+      int16_t a_command;
+      tftStrchrPtr = strchr(tftCommands[tftBufIndW], 'A');
+      a_command = ((int)((strtod(&tftCommands[tftBufIndW][tftStrchrPtr - tftCommands[tftBufIndW] + 1], nullptr))));
 
-        #if ENABLED(ANYCUBIC_LCD_DEBUG)
-          if ((a_command > 7) && (a_command != 20))   // No debugging of status polls, please!
-            SERIAL_ECHOLNPGM("TFT Serial Command: ", TFTcmdbuffer[TFTbufindw]);
-        #endif
-
-        switch (a_command) {
-          case 0: { // A0 GET HOTEND TEMP
-            const celsius_float_t hotendActualTemp = getActualTemp_celsius(E0);
-            SEND_PGM_VAL("A0V ", ROUND(hotendActualTemp));
-          }
-          break;
-
-          case 1: { // A1  GET HOTEND TARGET TEMP
-            const celsius_float_t hotendTargetTemp = getTargetTemp_celsius(E0);
-            SEND_PGM_VAL("A1V ", ROUND(hotendTargetTemp));
-          }
-          break;
-
-          case 2: { // A2 GET HOTBED TEMP
-            const celsius_float_t heatedBedActualTemp = getActualTemp_celsius(BED);
-            SEND_PGM_VAL("A2V ", ROUND(heatedBedActualTemp));
-          }
-          break;
-
-          case 3: { // A3 GET HOTBED TARGET TEMP
-            const celsius_float_t heatedBedTargetTemp = getTargetTemp_celsius(BED);
-            SEND_PGM_VAL("A3V ", ROUND(heatedBedTargetTemp));
-          } break;
-
-          case 4: { // A4 GET FAN SPEED
-            SEND_PGM_VAL("A4V ", int(getActualFan_percent(FAN0)));
-          } break;
-
-          case 5: { // A5 GET CURRENT COORDINATE
-            const float xPosition = getAxisPosition_mm(X),
-                        yPosition = getAxisPosition_mm(Y),
-                        zPosition = getAxisPosition_mm(Z);
-            SEND_PGM("A5V X: "); LCD_SERIAL.print(xPosition);
-            SEND_PGM(   " Y: "); LCD_SERIAL.print(yPosition);
-            SEND_PGM(   " Z: "); LCD_SERIAL.print(zPosition);
-            SENDLINE_PGM("");
-          } break;
-
-          case 6: // A6 GET SD CARD PRINTING STATUS
-            #if HAS_MEDIA
-              if (isPrintingFromMedia()) {
-                SEND_PGM("A6V ");
-                if (isMediaInserted())
-                  SENDLINE(ui8tostr3rj(getProgress_percent()));
-                else
-                  SENDLINE_DBG_PGM("J02", "TFT Serial Debug: No SD Card mounted to return printing status... J02");
-              }
-              else
-                SENDLINE_PGM("A6V ---");
-            #endif
-            break;
-
-          case 7: { // A7 GET PRINTING TIME
-            const uint32_t elapsedSeconds = getProgress_seconds_elapsed();
-            SEND_PGM("A7V ");
-            if (elapsedSeconds != 0) {  // print time
-              const uint32_t elapsedMinutes = elapsedSeconds / 60;
-              SEND(ui8tostr2(elapsedMinutes / 60));
-              SEND_PGM(" H ");
-              SEND(ui8tostr2(elapsedMinutes % 60));
-              SENDLINE_PGM(" M");
-            }
-            else
-              SENDLINE_PGM(" 999:999");
-          }
-          break;
-
-          case 8: // A8 GET  SD LIST
-            #if HAS_MEDIA
-              SelectedFile[0] = 0;
-              RenderCurrentFileList();
-            #endif
-            break;
-
-          case 9: // A9 pause sd print
-            #if HAS_MEDIA
-              if (isPrintingFromMedia())
-                PausePrint();
-            #endif
-            break;
-
-          case 10: // A10 resume sd print
-            #if HAS_MEDIA
-              if (isPrintingFromMediaPaused())
-                ResumePrint();
-            #endif
-            break;
-
-          case 11: // A11 STOP SD PRINT
-            TERN_(HAS_MEDIA, StopPrint());
-            break;
-
-          case 12: // A12 kill
-            kill(F(STR_ERR_KILLED));
-            break;
-
-          case 13: // A13 SELECTION FILE
-            #if HAS_MEDIA
-              if (isMediaInserted()) {
-                starpos = (strchr(TFTstrchr_pointer + 4, '*'));
-                if (TFTstrchr_pointer[4] == '/') {
-                  strcpy(SelectedDirectory, TFTstrchr_pointer + 5);
-                  SelectedFile[0] = 0;
-                  SENDLINE_DBG_PGM("J21", "TFT Serial Debug: Clear file selection... J21 "); // J21 Not File Selected
-                  SENDLINE_PGM("");
-                }
-                else if (TFTstrchr_pointer[4] == '<') {
-                  strcpy(SelectedDirectory, TFTstrchr_pointer + 4);
-                  SpecialMenu = true;
-                  SelectedFile[0] = 0;
-                  SENDLINE_DBG_PGM("J21", "TFT Serial Debug: Clear file selection... J21 "); // J21 Not File Selected
-                  SENDLINE_PGM("");
-                }
-                else {
-                  SelectedDirectory[0] = 0;
-
-                  if (starpos) *(starpos - 1) = '\0';
-
-                  strcpy(SelectedFile, TFTstrchr_pointer + 4);
-                  SENDLINE_DBG_PGM_VAL("J20", "TFT Serial Debug: File Selected... J20 ", SelectedFile); // J20 File Selected
-                }
-              }
-            #endif
-            break;
-
-          case 14: // A14 START PRINTING
-            #if HAS_MEDIA
-              if (!isPrinting() && strlen(SelectedFile) > 0)
-                StartPrint();
-            #endif
-            break;
-
-          case 15: // A15 RESUMING FROM OUTAGE
-            // TODO: JBA implement resume form outage
-            break;
-
-          case 16: { // A16 set hotend temp
-            uint16_t tempvalue;
-            if (CodeSeen('S')) {
-              tempvalue = constrain(CodeValue(), 0, 275);
-              setTargetTemp_celsius(tempvalue, (extruder_t)E0);
-            }
-            else if (CodeSeen('C') && !isPrinting()) {
-              if (getAxisPosition_mm(Z) < 10)
-                injectCommands(F("G1 Z10")); // RASE Z AXIS
-              tempvalue = constrain(CodeValue(), 0, 275);
-              setTargetTemp_celsius(tempvalue, (extruder_t)E0);
-            }
-          }
-          break;
-
-          case 17: { // A17 set heated bed temp
-            uint16_t tempbed;
-            if (CodeSeen('S')) {
-              tempbed = constrain(CodeValue(), 0, 100);
-              setTargetTemp_celsius(tempbed, (heater_t)BED);
-            }
-          }
-          break;
-
-          case 18: { // A18 set fan speed
-            float fanPercent;
-            if (CodeSeen('S')) {
-              fanPercent = CodeValue();
-              fanPercent = constrain(fanPercent, 0, 100);
-              setTargetFan_percent(fanPercent, FAN0);
-            }
-            else
-              fanPercent = 100;
-
-            setTargetFan_percent(fanPercent, FAN0);
-            SENDLINE_PGM("");
-          }
-          break;
-
-          case 19: // A19 stop stepper drivers - sent on stop extrude command and on turn motors off command
-            if (!isPrinting()) {
-              quickstop_stepper();
-              stepper.disable_all_steppers();
-            }
-
-            SENDLINE_PGM("");
-            break;
-
-          case 20: // A20 read printing speed
-            if (CodeSeen('S'))
-              feedrate_percentage = constrain(CodeValue(), 40, 999);
-            else
-              SEND_PGM_VAL("A20V ", feedrate_percentage);
-            break;
-
-          case 21: // A21 all home
-            if (!isPrinting() && !isPrintingFromMediaPaused()) {
-              if (CodeSeen('X') || CodeSeen('Y') || CodeSeen('Z')) {
-                if (CodeSeen('X'))
-                  injectCommands(F("G28X"));
-                if (CodeSeen('Y'))
-                  injectCommands(F("G28Y"));
-                if (CodeSeen('Z'))
-                  injectCommands(F("G28Z"));
-              }
-              else if (CodeSeen('C')) {
-                injectCommands_P(G28_STR);
-              }
-            }
-            break;
-
-          case 22: // A22 move X/Y/Z or extrude
-            if (!isPrinting()) {
-              float coorvalue;
-              uint16_t movespeed = 0;
-              char commandStr[30];
-              char fullCommandStr[38];
-
-              commandStr[0] = 0; // empty string
-              if (CodeSeen('F'))  // Set feedrate
-                movespeed = CodeValue();
-
-              if (CodeSeen('X')) { // Move in X direction
-                coorvalue = CodeValue();
-                if ((coorvalue <= 0.2) && coorvalue > 0)
-                  sprintf_P(commandStr, PSTR("G1 X0.1F%i"), movespeed);
-                else if ((coorvalue <= -0.1) && coorvalue > -1)
-                  sprintf_P(commandStr, PSTR("G1 X-0.1F%i"), movespeed);
-                else
-                  sprintf_P(commandStr, PSTR("G1 X%iF%i"), int(coorvalue), movespeed);
-              }
-              else if (CodeSeen('Y')) {  // Move in Y direction
-                coorvalue = CodeValue();
-                if ((coorvalue <= 0.2) && coorvalue > 0)
-                  sprintf_P(commandStr, PSTR("G1 Y0.1F%i"), movespeed);
-                else if ((coorvalue <= -0.1) && coorvalue > -1)
-                  sprintf_P(commandStr, PSTR("G1 Y-0.1F%i"), movespeed);
-                else
-                  sprintf_P(commandStr, PSTR("G1 Y%iF%i"), int(coorvalue), movespeed);
-              }
-              else if (CodeSeen('Z')) {  // Move in Z direction
-                coorvalue = CodeValue();
-                if ((coorvalue <= 0.2) && coorvalue > 0)
-                  sprintf_P(commandStr, PSTR("G1 Z0.1F%i"), movespeed);
-                else if ((coorvalue <= -0.1) && coorvalue > -1)
-                  sprintf_P(commandStr, PSTR("G1 Z-0.1F%i"), movespeed);
-                else
-                  sprintf_P(commandStr, PSTR("G1 Z%iF%i"), int(coorvalue), movespeed);
-              }
-              else if (CodeSeen('E')) { // Extrude
-                coorvalue = CodeValue();
-                if ((coorvalue <= 0.2) && coorvalue > 0)
-                  sprintf_P(commandStr, PSTR("G1 E0.1F%i"), movespeed);
-                else if ((coorvalue <= -0.1) && coorvalue > -1)
-                  sprintf_P(commandStr, PSTR("G1 E-0.1F%i"), movespeed);
-                else
-                  sprintf_P(commandStr, PSTR("G1 E%iF500"), int(coorvalue));
-              }
-
-              if (strlen(commandStr) > 0) {
-                sprintf_P(fullCommandStr, PSTR("G91\n%s\nG90"), commandStr);
-                #if ENABLED(ANYCUBIC_LCD_DEBUG)
-                  SERIAL_ECHOPGM("TFT Serial Debug: A22 Move final request with gcode... ");
-                  SERIAL_ECHOLN(fullCommandStr);
-                #endif
-                injectCommands(fullCommandStr);
-              }
-            }
-            SENDLINE_PGM("");
-            break;
-
-          case 23: // A23 preheat pla
-            if (!isPrinting()) {
-              if (getAxisPosition_mm(Z) < 10)
-                injectCommands(F("G1 Z10")); // RASE Z AXIS
-
-              setTargetTemp_celsius(PREHEAT_1_TEMP_BED, (heater_t)BED);
-              setTargetTemp_celsius(PREHEAT_1_TEMP_HOTEND, (extruder_t)E0);
-              SENDLINE_PGM("OK");
-            }
-            break;
-
-          case 24:// A24 preheat abs
-            if (!isPrinting()) {
-              if (getAxisPosition_mm(Z) < 10)
-                injectCommands(F("G1 Z10")); // RASE Z AXIS
-
-              setTargetTemp_celsius(PREHEAT_2_TEMP_BED, (heater_t)BED);
-              setTargetTemp_celsius(PREHEAT_2_TEMP_HOTEND, (extruder_t)E0);
-              SENDLINE_PGM("OK");
-            }
-            break;
-
-          case 25: // A25 cool down
-            if (!isPrinting()) {
-              setTargetTemp_celsius(0, (heater_t) BED);
-              setTargetTemp_celsius(0, (extruder_t) E0);
-
-              SENDLINE_DBG_PGM("J12", "TFT Serial Debug: Cooling down... J12"); // J12 cool down
-            }
-            break;
-
-          case 26: // A26 refresh SD
-            #if HAS_MEDIA
-              if (isMediaInserted()) {
-                if (strlen(SelectedDirectory) > 0) {
-                  FileList currentFileList;
-                  if ((SelectedDirectory[0] == '.') && (SelectedDirectory[1] == '.')) {
-                    currentFileList.upDir();
-                  }
-                  else {
-                    if (SelectedDirectory[0] == '<')
-                      HandleSpecialMenu();
-                    else
-                      currentFileList.changeDir(SelectedDirectory);
-                  }
-                }
-              }
-              else {
-                SENDLINE_DBG_PGM("J02", "TFT Serial Debug: No SD Card mounted to refresh SD A26... J02");
-              }
-
-              SelectedDirectory[0] = 0;
-            #endif
-            break;
-
-            #if ENABLED(SERVO_ENDSTOPS)
-              case 27: break; // A27 servos angles adjust
-            #endif
-
-          case 28: // A28 filament test
-            if (CodeSeen('O'))
-              NOOP;
-            else if (CodeSeen('C'))
-              NOOP;
-            SENDLINE_PGM("");
-            break;
-
-          case 33: // A33 get version info
-            SEND_PGM("J33 ");
-            SENDLINE_PGM(DETAILED_BUILD_VERSION);
-            break;
-
-          default:
-            break;
-        }
+      if (a_command > 7 && a_command != 20) { // No debugging of status polls, please!
+        DEBUG_ECHOLNPGM("TFT Serial Command: ", tftCommands[tftBufIndW]);
       }
 
-      TFTbufindw = (TFTbufindw + 1) % TFTBUFSIZE;
-      TFTbuflen += 1;
-      serial3_count = 0; // clear buffer
+      switch (a_command) {
+        case 0: { // A0 GET HOTEND TEMP
+          const celsius_float_t hotendActualTemp = getActualTemp_celsius(E0);
+          SEND_PGM_VAL("A0V ", ROUND(hotendActualTemp));
+        }
+        break;
+
+        case 1: { // A1  GET HOTEND TARGET TEMP
+          const celsius_float_t hotendTargetTemp = getTargetTemp_celsius(E0);
+          SEND_PGM_VAL("A1V ", ROUND(hotendTargetTemp));
+        }
+        break;
+
+        case 2: { // A2 GET HOTBED TEMP
+          const celsius_float_t heatedBedActualTemp = getActualTemp_celsius(BED);
+          SEND_PGM_VAL("A2V ", ROUND(heatedBedActualTemp));
+        }
+        break;
+
+        case 3: { // A3 GET HOTBED TARGET TEMP
+          const celsius_float_t heatedBedTargetTemp = getTargetTemp_celsius(BED);
+          SEND_PGM_VAL("A3V ", ROUND(heatedBedTargetTemp));
+        } break;
+
+        case 4: { // A4 GET FAN SPEED
+          SEND_PGM_VAL("A4V ", int(getActualFan_percent(FAN0)));
+        } break;
+
+        case 5: { // A5 GET CURRENT COORDINATE
+          const float xPosition = getAxisPosition_mm(X),
+                      yPosition = getAxisPosition_mm(Y),
+                      zPosition = getAxisPosition_mm(Z);
+          SEND_PGM("A5V X: "); LCD_SERIAL.print(xPosition);
+          SEND_PGM(   " Y: "); LCD_SERIAL.print(yPosition);
+          SEND_PGM(   " Z: "); LCD_SERIAL.print(zPosition);
+          SENDLINE_PGM("");
+        } break;
+
+        case 6: // A6 GET SD CARD PRINTING STATUS
+          #if HAS_MEDIA
+            if (isPrintingFromMedia()) {
+              SEND_PGM("A6V ");
+              if (isMediaInserted())
+                SENDLINE(ui8tostr3rj(getProgress_percent()));
+              else
+                SENDLINE_DBG_PGM("J02", "TFT Serial Debug: No SD Card mounted to return printing status... J02");
+            }
+            else
+              SENDLINE_PGM("A6V ---");
+          #endif
+          break;
+
+        case 7: { // A7 GET PRINTING TIME
+          const uint32_t elapsedSeconds = getProgress_seconds_elapsed();
+          SEND_PGM("A7V ");
+          if (elapsedSeconds != 0) {  // print time
+            const uint32_t elapsedMinutes = elapsedSeconds / 60;
+            SEND(ui8tostr2(elapsedMinutes / 60));
+            SEND_PGM(" H ");
+            SEND(ui8tostr2(elapsedMinutes % 60));
+            SENDLINE_PGM(" M");
+          }
+          else
+            SENDLINE_PGM(" 999:999");
+        }
+        break;
+
+        case 8: // A8 GET  SD LIST
+          #if HAS_MEDIA
+            selectedFile[0] = 0;
+            renderCurrentFileList();
+          #endif
+          break;
+
+        case 9: // A9 pause sd print
+          TERN_(HAS_MEDIA, if (isPrintingFromMedia()) pausePrint());
+          break;
+
+        case 10: // A10 resume sd print
+          #if HAS_MEDIA
+            if (isPrintingFromMediaPaused())
+              resumePrint();
+          #endif
+          break;
+
+        case 11: // A11 STOP SD PRINT
+          TERN_(HAS_MEDIA, stopPrint());
+          break;
+
+        case 12: // A12 kill
+          kill(F(STR_ERR_KILLED));
+          break;
+
+        case 13: // A13 SELECTION FILE
+          #if HAS_MEDIA
+            if (isMediaInserted()) {
+              starpos = (strchr(tftStrchrPtr + 4, '*'));
+              if (tftStrchrPtr[4] == '/') {
+                strcpy(selectedDirectory, tftStrchrPtr + 5);
+                selectedFile[0] = 0;
+                SENDLINE_DBG_PGM("J21", "TFT Serial Debug: Clear file selection... J21 "); // J21 Not File Selected
+                SENDLINE_PGM("");
+              }
+              else if (tftStrchrPtr[4] == '<') {
+                strcpy(selectedDirectory, tftStrchrPtr + 4);
+                specialMenu = true;
+                selectedFile[0] = 0;
+                SENDLINE_DBG_PGM("J21", "TFT Serial Debug: Clear file selection... J21 "); // J21 Not File Selected
+                SENDLINE_PGM("");
+              }
+              else {
+                selectedDirectory[0] = 0;
+
+                if (starpos) *(starpos - 1) = '\0';
+
+                strcpy(selectedFile, tftStrchrPtr + 4);
+                SENDLINE_DBG_PGM_VAL("J20", "TFT Serial Debug: File Selected... J20 ", selectedFile); // J20 File Selected
+              }
+            }
+          #endif
+          break;
+
+        case 14: // A14 START PRINTING
+          #if HAS_MEDIA
+            if (!isPrinting() && strlen(selectedFile) > 0)
+              startPrint();
+          #endif
+          break;
+
+        case 15: // A15 RESUMING FROM OUTAGE
+          // TODO: JBA implement resume form outage
+          break;
+
+        case 16: { // A16 set hotend temp
+          uint16_t tempvalue;
+          if (codeSeen('S')) {
+            tempvalue = constrain(codeValue(), 0, 275);
+            setTargetTemp_celsius(tempvalue, (extruder_t)E0);
+          }
+          else if (codeSeen('C') && !isPrinting()) {
+            if (getAxisPosition_mm(Z) < 10)
+              injectCommands(F("G1 Z10")); // RASE Z AXIS
+            tempvalue = constrain(codeValue(), 0, 275);
+            setTargetTemp_celsius(tempvalue, (extruder_t)E0);
+          }
+        }
+        break;
+
+        case 17: { // A17 set heated bed temp
+          uint16_t tempbed;
+          if (codeSeen('S')) {
+            tempbed = constrain(codeValue(), 0, 100);
+            setTargetTemp_celsius(tempbed, (heater_t)BED);
+          }
+        }
+        break;
+
+        case 18: { // A18 set fan speed
+          float fanPercent;
+          if (codeSeen('S')) {
+            fanPercent = codeValue();
+            fanPercent = constrain(fanPercent, 0, 100);
+            setTargetFan_percent(fanPercent, FAN0);
+          }
+          else
+            fanPercent = 100;
+
+          setTargetFan_percent(fanPercent, FAN0);
+          SENDLINE_PGM("");
+        }
+        break;
+
+        case 19: // A19 stop stepper drivers - sent on stop extrude command and on turn motors off command
+          if (!isPrinting()) {
+            quickstop_stepper();
+            stepper.disable_all_steppers();
+          }
+
+          SENDLINE_PGM("");
+          break;
+
+        case 20: // A20 read printing speed
+          if (codeSeen('S'))
+            feedrate_percentage = constrain(codeValue(), 40, 999);
+          else
+            SEND_PGM_VAL("A20V ", feedrate_percentage);
+          break;
+
+        case 21: // A21 all home
+          if (!isPrinting() && !isPrintingFromMediaPaused()) {
+            if (codeSeen('X') || codeSeen('Y') || codeSeen('Z')) {
+              if (codeSeen('X'))
+                injectCommands(F("G28X"));
+              if (codeSeen('Y'))
+                injectCommands(F("G28Y"));
+              if (codeSeen('Z'))
+                injectCommands(F("G28Z"));
+            }
+            else if (codeSeen('C')) {
+              injectCommands_P(G28_STR);
+            }
+          }
+          break;
+
+        case 22: // A22 move X/Y/Z or extrude
+          if (!isPrinting()) {
+            float coorvalue;
+            uint16_t movespeed = 0;
+            char commandStr[30];
+            char fullCommandStr[38];
+
+            commandStr[0] = 0; // empty string
+            if (codeSeen('F'))  // Set feedrate
+              movespeed = codeValue();
+
+            if (codeSeen('X')) { // Move in X direction
+              coorvalue = codeValue();
+              if ((coorvalue <= 0.2) && coorvalue > 0)
+                sprintf_P(commandStr, PSTR("G1 X0.1F%i"), movespeed);
+              else if ((coorvalue <= -0.1) && coorvalue > -1)
+                sprintf_P(commandStr, PSTR("G1 X-0.1F%i"), movespeed);
+              else
+                sprintf_P(commandStr, PSTR("G1 X%iF%i"), int(coorvalue), movespeed);
+            }
+            else if (codeSeen('Y')) {  // Move in Y direction
+              coorvalue = codeValue();
+              if ((coorvalue <= 0.2) && coorvalue > 0)
+                sprintf_P(commandStr, PSTR("G1 Y0.1F%i"), movespeed);
+              else if ((coorvalue <= -0.1) && coorvalue > -1)
+                sprintf_P(commandStr, PSTR("G1 Y-0.1F%i"), movespeed);
+              else
+                sprintf_P(commandStr, PSTR("G1 Y%iF%i"), int(coorvalue), movespeed);
+            }
+            else if (codeSeen('Z')) {  // Move in Z direction
+              coorvalue = codeValue();
+              if ((coorvalue <= 0.2) && coorvalue > 0)
+                sprintf_P(commandStr, PSTR("G1 Z0.1F%i"), movespeed);
+              else if ((coorvalue <= -0.1) && coorvalue > -1)
+                sprintf_P(commandStr, PSTR("G1 Z-0.1F%i"), movespeed);
+              else
+                sprintf_P(commandStr, PSTR("G1 Z%iF%i"), int(coorvalue), movespeed);
+            }
+            else if (codeSeen('E')) { // Extrude
+              coorvalue = codeValue();
+              if ((coorvalue <= 0.2) && coorvalue > 0)
+                sprintf_P(commandStr, PSTR("G1 E0.1F%i"), movespeed);
+              else if ((coorvalue <= -0.1) && coorvalue > -1)
+                sprintf_P(commandStr, PSTR("G1 E-0.1F%i"), movespeed);
+              else
+                sprintf_P(commandStr, PSTR("G1 E%iF500"), int(coorvalue));
+            }
+
+            if (strlen(commandStr) > 0) {
+              sprintf_P(fullCommandStr, PSTR("G91\n%s\nG90"), commandStr);
+              DEBUG_ECHOPGM("TFT Serial Debug: A22 Move final request with gcode... ");
+              DEBUG_ECHOLN(fullCommandStr);
+              injectCommands(fullCommandStr);
+            }
+          }
+          SENDLINE_PGM("");
+          break;
+
+        case 23: // A23 preheat pla
+          if (!isPrinting()) {
+            if (getAxisPosition_mm(Z) < 10)
+              injectCommands(F("G1 Z10")); // RASE Z AXIS
+
+            setTargetTemp_celsius(PREHEAT_1_TEMP_BED, (heater_t)BED);
+            setTargetTemp_celsius(PREHEAT_1_TEMP_HOTEND, (extruder_t)E0);
+            SENDLINE_PGM("OK");
+          }
+          break;
+
+        case 24:// A24 preheat abs
+          if (!isPrinting()) {
+            if (getAxisPosition_mm(Z) < 10)
+              injectCommands(F("G1 Z10")); // RASE Z AXIS
+
+            setTargetTemp_celsius(PREHEAT_2_TEMP_BED, (heater_t)BED);
+            setTargetTemp_celsius(PREHEAT_2_TEMP_HOTEND, (extruder_t)E0);
+            SENDLINE_PGM("OK");
+          }
+          break;
+
+        case 25: // A25 cool down
+          if (!isPrinting()) {
+            setTargetTemp_celsius(0, (heater_t) BED);
+            setTargetTemp_celsius(0, (extruder_t) E0);
+
+            SENDLINE_DBG_PGM("J12", "TFT Serial Debug: Cooling down... J12"); // J12 cool down
+          }
+          break;
+
+        case 26: // A26 refresh SD
+          #if HAS_MEDIA
+            if (isMediaInserted()) {
+              if (strlen(selectedDirectory) > 0) {
+                FileList currentFileList;
+                if ((selectedDirectory[0] == '.') && (selectedDirectory[1] == '.')) {
+                  currentFileList.upDir();
+                }
+                else {
+                  if (selectedDirectory[0] == '<')
+                    handleSpecialMenu();
+                  else
+                    currentFileList.changeDir(selectedDirectory);
+                }
+              }
+            }
+            else {
+              SENDLINE_DBG_PGM("J02", "TFT Serial Debug: No SD Card mounted to refresh SD A26... J02");
+            }
+
+            selectedDirectory[0] = 0;
+          #endif
+          break;
+
+          #if ENABLED(SERVO_ENDSTOPS)
+            case 27: break; // A27 servos angles adjust
+          #endif
+
+        case 28: // A28 filament test
+          if (codeSeen('O'))
+            NOOP;
+          else if (codeSeen('C'))
+            NOOP;
+          SENDLINE_PGM("");
+          break;
+
+        case 33: // A33 get version info
+          SEND_PGM("J33 ");
+          SENDLINE_PGM(DETAILED_BUILD_VERSION);
+          break;
+
+        default:
+          break;
+      }
     }
-    else {
-      TFTcmdbuffer[TFTbufindw][serial3_count++] = serial3_char;
-    }
-  }
+
+    tftBufIndW = (tftBufIndW + 1) % (TFTBUFSIZE);
+    tftBufLen++;
+    serial_count = 0; // clear buffer
+
+  } // while
 }
 
-void AnycubicTFTClass::DoSDCardStateCheck() {
+void AnycubicTFTClass::doSDCardStateCheck() {
   #if BOTH(HAS_MEDIA, HAS_SD_DETECT)
     bool isInserted = isMediaInserted();
     if (isInserted)
@@ -934,7 +895,7 @@ void AnycubicTFTClass::DoSDCardStateCheck() {
   #endif
 }
 
-void AnycubicTFTClass::DoFilamentRunoutCheck() {
+void AnycubicTFTClass::doFilamentRunoutCheck() {
   #if ENABLED(FILAMENT_RUNOUT_SENSOR)
     // NOTE: getFilamentRunoutState() only returns the runout state if the job is printing
     // we want to actually check the status of the pin here, regardless of printstate
@@ -953,23 +914,18 @@ void AnycubicTFTClass::DoFilamentRunoutCheck() {
   #endif // FILAMENT_RUNOUT_SENSOR
 }
 
-void AnycubicTFTClass::StartPrint() {
+void AnycubicTFTClass::startPrint() {
   #if HAS_MEDIA
-    if (!isPrinting() && strlen(SelectedFile) > 0) {
-      #if ENABLED(ANYCUBIC_LCD_DEBUG)
-        SERIAL_ECHOPGM("TFT Serial Debug: About to print file ... ");
-        SERIAL_ECHO(isPrinting());
-        SERIAL_ECHOPGM(" ");
-        SERIAL_ECHOLN(SelectedFile);
-      #endif
+    if (!isPrinting() && strlen(selectedFile) > 0) {
+      DEBUG_ECHOLNPGM("TFT Serial Debug: About to print file ... ", isPrinting(), " ", selectedFile);
       mediaPrintingState = AMPRINTSTATE_PRINTING;
       mediaPauseState    = AMPAUSESTATE_NOT_PAUSED;
-      printFile(SelectedFile);
+      printFile(selectedFile);
     }
   #endif // SDUPPORT
 }
 
-void AnycubicTFTClass::PausePrint() {
+void AnycubicTFTClass::pausePrint() {
   #if HAS_MEDIA
     if (isPrintingFromMedia() && mediaPrintingState != AMPRINTSTATE_STOP_REQUESTED && mediaPauseState == AMPAUSESTATE_NOT_PAUSED) {
       mediaPrintingState = AMPRINTSTATE_PAUSE_REQUESTED;
@@ -983,16 +939,14 @@ void AnycubicTFTClass::PausePrint() {
   #endif
 }
 
-void AnycubicTFTClass::ResumePrint() {
+void AnycubicTFTClass::resumePrint() {
   #if HAS_MEDIA
     #if ENABLED(FILAMENT_RUNOUT_SENSOR)
       if (READ(FIL_RUNOUT1_PIN)) {
-        #if ENABLED(ANYCUBIC_LCD_DEBUG)
-          SERIAL_ECHOLNPGM("TFT Serial Debug: Resume Print with filament sensor still tripped... ");
-        #endif
+        DEBUG_ECHOLNPGM("TFT Serial Debug: Resume Print with filament sensor still tripped... ");
 
         // trigger the user message box
-        DoFilamentRunoutCheck();
+        doFilamentRunoutCheck();
 
         // re-enable the continue button
         SENDLINE_DBG_PGM("J18", "TFT Serial Debug: Resume Print with filament sensor still tripped... J18");
@@ -1019,7 +973,7 @@ void AnycubicTFTClass::ResumePrint() {
   #endif
 }
 
-void AnycubicTFTClass::StopPrint() {
+void AnycubicTFTClass::stopPrint() {
   #if HAS_MEDIA
     mediaPrintingState = AMPRINTSTATE_STOP_REQUESTED;
     mediaPauseState    = AMPAUSESTATE_NOT_PAUSED;
