+++ Marlin/src/lcd/extui/anycubic_vyper/FileNavigator.cpp
@@ -40,8 +40,8 @@ using namespace ExtUI;
 
 namespace Anycubic {
 
-  FileList  FileNavigator::filelist;                            // Instance of the Marlin file API
-  char      FileNavigator::currentfoldername[MAX_PATH_LEN + 1]; // Current folder path
+  FileList  FileNavigator::filelist;                          // ExtUI file API
+  char      FileNavigator::currentDirPath[MAX_PATH_LEN + 1];  // Current folder path
   uint16_t  FileNavigator::lastindex;
   uint8_t   FileNavigator::folderdepth;
   uint16_t  FileNavigator::currentindex;                      // override the panel request
@@ -51,7 +51,7 @@ namespace Anycubic {
   FileNavigator::FileNavigator() { reset(); }
 
   void FileNavigator::reset() {
-    currentfoldername[0] = '\0';
+    currentDirPath[0] = '\0';
     folderdepth  = 0;
     currentindex = 0;
     lastindex    = 0;
@@ -94,7 +94,7 @@ namespace Anycubic {
         DgusTFT::SendTxtToTFT(filelist.longFilename(), TXT_FILE_0 + file_num*0x30);
 
         #if ACDEBUG(AC_FILE)
-          SERIAL_ECHOLNPGM("seek: ", _seek, " '", filelist.longFilename(), "' '", currentfoldername, "", filelist.shortFilename(), "'\n");
+          SERIAL_ECHOLNPGM("seek: ", _seek, " '", filelist.longFilename(), "' '", currentDirPath, "", filelist.shortFilename(), "'\n");
         #endif
       }
       else {
@@ -120,11 +120,11 @@ namespace Anycubic {
 
   void FileNavigator::changeDIR(char *folder) {
     #if ACDEBUG(AC_FILE)
-      SERIAL_ECHOLNPGM("currentfolder: ", currentfoldername, "  New: ", folder);
+      SERIAL_ECHOLNPGM("currentfolder: ", currentDirPath, "  New: ", folder);
     #endif
     if (folderdepth >= MAX_FOLDER_DEPTH) return; // limit the folder depth
-    strcat(currentfoldername, folder);
-    strcat(currentfoldername, "/");
+    strcat(currentDirPath, folder);
+    strcat(currentDirPath, "/");
     filelist.changeDir(folder);
     refresh();
     folderdepth++;
@@ -138,22 +138,22 @@ namespace Anycubic {
     currentindex = 0;
     // Remove the last child folder from the stored path
     if (folderdepth == 0) {
-      currentfoldername[0] = '\0';
+      currentDirPath[0] = '\0';
       reset();
     }
     else {
       char *pos = nullptr;
       for (uint8_t f = 0; f < folderdepth; f++)
-        pos = strchr(currentfoldername, '/');
+        pos = strchr(currentDirPath, '/');
 
       *(pos + 1) = '\0';
     }
     #if ACDEBUG(AC_FILE)
-      SERIAL_ECHOLNPGM("depth: ", folderdepth, " currentfoldername: ", currentfoldername);
+      SERIAL_ECHOLNPGM("depth: ", folderdepth, " currentDirPath: ", currentDirPath);
     #endif
   }
 
-  char* FileNavigator::getCurrentFolderName() { return currentfoldername; }
+  char* FileNavigator::getCurrentDirPath() { return currentDirPath; }
 
   uint16_t FileNavigator::getFileNum() { return filelist.count(); }
 }
