+++ Marlin/src/lcd/extui/dgus/mks/DGUSScreenHandler.cpp
@@ -739,7 +739,7 @@ void DGUSScreenHandler::HandleManualMove(DGUS_VP_Variable &var, void *val_ptr) {
     return;
 
   char axiscode;
-  uint16_t speed = manual_feedrate_mm_m[X_AXIS]; // Default feedrate for manual moves
+  uint16_t speed = manual_feedrate_mm_m.x; // Default feedrate for manual moves
 
   switch (var.VP) { // switch X Y Z or Home
     default: return;
@@ -753,7 +753,7 @@ void DGUSScreenHandler::HandleManualMove(DGUS_VP_Variable &var, void *val_ptr) {
     #if HAS_Y_AXIS
       case VP_MOVE_Y:
         axiscode = 'Y';
-        speed = manual_feedrate_mm_m[Y_AXIS];
+        speed = manual_feedrate_mm_m.y;
         if (!ExtUI::canMove(ExtUI::axis_t::Y)) goto cannotmove;
         break;
     #endif
@@ -761,7 +761,7 @@ void DGUSScreenHandler::HandleManualMove(DGUS_VP_Variable &var, void *val_ptr) {
     #if HAS_Z_AXIS
       case VP_MOVE_Z:
         axiscode = 'Z';
-        speed = manual_feedrate_mm_m[Z_AXIS];
+        speed = manual_feedrate_mm_m.z;
         if (!ExtUI::canMove(ExtUI::axis_t::Z)) goto cannotmove;
         break;
     #endif
