+++ Marlin/src/lcd/extui/dgus_reloaded/DGUSScreenHandler.h
@@ -37,7 +37,7 @@ public:
   static void Ready();
   static void Loop();
 
-  static void PrinterKilled(FSTR_P const error, FSTR_P const component);
+  static void printerKilled(FSTR_P const error, FSTR_P const component);
   static void UserConfirmRequired(const char * const msg);
   static void SettingsReset();
   static void StoreSettings(char *buff);
@@ -50,7 +50,7 @@ public:
   static void PrintTimerStarted();
   static void PrintTimerPaused();
   static void PrintTimerStopped();
-  static void FilamentRunout(const ExtUI::extruder_t extruder);
+  static void filamentRunout(const ExtUI::extruder_t extruder);
 
   #if HAS_MEDIA
     /// Marlin informed us that a new SD has been inserted.
