+++ Marlin/src/lcd/extui/ia_creality/FileNavigator.h
@@ -30,12 +30,11 @@
  * Written by Insanity Automation
  * ***************************************/
 
+#include "../ui_api.h"
+
 #define MAX_FOLDER_DEPTH   4    // Limit folder depth TFT has a limit for the file path
-#define MAX_CMND_LEN      16 * MAX_FOLDER_DEPTH // Maximum Length for a Panel command
 #define MAX_PATH_LEN      16 * MAX_FOLDER_DEPTH // Maximum number of characters in a SD file path
 #define DISPLAY_FILES      4
-
-using namespace ExtUI;
 
 class FileNavigator {
   public:
@@ -45,15 +44,15 @@ class FileNavigator {
     static void upDIR();
     static void changeDIR(char *);
     static void refresh();
-    static char* getCurrentFolderName();
+    static char* getCurrentDirPath();
     static uint8_t  folderdepth;
     static uint16_t currentindex;
     static bool getIndexisDir(uint16_t);
     const char *getIndexName(uint16_t);
     static uint16_t maxFiles();
   private:
-    static FileList filelist;
-    static char     currentfoldername[MAX_PATH_LEN];
+    static ExtUI::FileList filelist;
+    static char     currentDirPath[MAX_PATH_LEN];
     static uint16_t lastindex;
 };
