+++ Marlin/src/lcd/extui/nextion/FileNavigator.h
@@ -41,10 +41,10 @@ class FileNavigator {
     static void  upDIR();
     static void  changeDIR(char *);
     static void  refresh();
-    static char* getCurrentFolderName();
+    static char* getCurrentDirPath();
   private:
     static FileList filelist;
-    static char     currentfoldername[MAX_PATH_LEN];
+    static char     currentDirPath[MAX_PATH_LEN];
     static uint16_t lastindex;
     static uint8_t  folderdepth;
     static uint16_t currentindex;
